app: Add debug options to page footer in development mode
This adds the ability for controllers to only be loaded within a given
application environment, defined in controllers.yml as such:

    path: /-/debug
    controller: DashDebugController
    only:
      env:
        - development
        - test

A single application environment can be specified, instead of a list:

    only:
      env: development

The added DashDebugController allows for restarting the application
server, and reloading the language files. Links have been added to the
page footer (in partials/footer_debug_options.haml, rendered only if
the app is in development mode) to perform these actions.