models: Add initial case profile migration and models
This includes Case, CaseNote, CaseSpend, and CaseSpendYear.

CaseSpendYear is a searchable aggregate of CaseSpend entries by year
for each individual case, which will be regenerated every time a
CaseSpend is created. This means we can quickly get the total of this
year's spend for a case to display on their case profile page.
The following will get you this year's spending for the case `case_id`:

    Kukupa::Models::CaseSpendYear
      .get_case_year(case_id, DateTime.now.strftime('%Y'))
      .decrypt(:amount)