case/note: add `metadata` column
This is used for outside request author name/contact information, but
may have other uses in the future. Metadata on a CaseNote is parsed as
JSON put in the renderable object by `get_renderables`.

In the case of outside requests, the `name`, `phone`, and `email` keys
of the metadata JSON are displayed in the UI above the request content.