case/spend/edit: Allow editing and deleting case spends
If a case spend is edited, the spend amount is below the auto-approval
threshold, and the maximum spend per case-year has not been met (with
this edit added to the total), the edited spend will be approved
automatically.

If the edited spend is over the auto-approval threshold, it will have
it's approval revoked, even if the user editing the spend has approval
powers.