case/index: Add overall spending stats and per-case details
This extracts the case info generation from CaseIndexController into the
CaseIndexHelpers helper module. It has been improved to only grab each
advocate's details once (when they're first encountered while iterating
through the cases), and the rejection of not-assigned-to-me cases for
non-admins is done earlier, making it faster.

This also adds a list underneath each case, containing:

- The last case note's creation ts and author
- The last approved spend's creation ts, author, and approver
- The last unapproved spend's creation ts and author
- The total spending for the case for this month
- The total spending for the case for this year (incl. percentage of max)