helpers/case: Move user gathering from CaseIndexHelpers to CaseHelpers
This renames `case_index_populate_advocate` to `case_populate_advocate`
and adds the `:me` key, indicating whether that user is the currently
logged in user, to the individual user descriptors.

This also makes `CaseViewHelpers::get_renderables` use the renamed
function, which will speed things up a bit in real workloads (as it's
not retrieving the user information every time for every individual
case note / spend).