case/task: Don't encrypt CaseTaskUpdate type
This changes all handling of CaseTaskUpdate#update_type to use plain
text values, rather than encrypting. This also brings the JSON
transformation of CaseTaskUpdate entries in the renderable generation in
line with the other renderable types that use JSON.

This commit introduces a temporary addition to FilterRefreshWorker to
decrypt the contents of CaseTaskUpdate#update_type if the content of
that value is not one of the known update types. This should be removed
before releasing a production version of Kūkupa, it's only useful if
you happen to have a database with CaseTaskUpdates created before this
commit.