case/view: Split renderable generation up into models, allow pagination
This makes each type of renderable be generated on the models themselves,
meaning that we can just get a list of classes and IDs in the
CaseViewHelpers#get_renderables method. This allows us to paginate the
list of IDs, and then only make the database queries for the ones that
are actually being rendered.

Doing this makes everything a lot cleaner, at least in my opinion.

Pagination is not implemented in the frontend yet.