case/sync: Add generic case sync worker
This adds a `reconnect_last_sync` column to cases, and adds the
"reconnect-sync-after" configuration key. If the worker is called, and
the time in the last sync column is older than the parsed value of the
sync configuration key, the sync is performed.

There will soon be a worker that calls this for every case.