case/correspondence: Add initial support for correspondence templates
These are stored in the database as encrypted HTML blobs, and when a
template ID is passed to the composer controller as `?tpl=<ID>`, the
template name is populated in the subject field, and the template
content is populated in the composer.

No way to create templates exists within the admin user interface yet,
but they can be manually created via the console.

See againstprisons/kukupa#10.