case: Allow multiple assigned advocates for a case
This adds a CaseAssignedAdvocate database table, which has foreign keys
to a case and a user. The Case#get_assigned_advocates method has been
updated to get the user IDs from that table, and the Case::assigned_to
method has been updated to get the cases from that table for a given
user.

CaseIndexController and it's associated view has been updated to show
the list of all advocates assigned to the cases that are assigned to
the current user.

CaseEditController has been updated to allow adding an advocate to the
list of assigned advocates, as well as allowing for removing an
advocate from the list of assignees. Roles have been updated for both
these actions - `case:assignees:assign` for assigning a new advocate,
and `case:assignees:unassign` for removing one.

CaseHelpers::case_users_with_access has been updated to use the list of
assigned advocates for the case as the starting point.

CaseHelpers::case_user_can_access? has been updated to check the list
of assigned advocates for the case.

The email alert methods on CaseNote and CaseSpend have been updated to
send emails to the full list of advocates for the case.

The database migration that enables this change creates the new
CaseAssignedAdvocate table and creates an entry in it for the assigned
advocate for each case, before dropping the `assigned_advocate` column
of the Case table. The migration file is commented pretty well with the
details of the changes, which may help if something goes wrong.

Please, for the love of God, take a database backup before upgrading to
this commit. I really mean it.

Closes againstprisons/kukupa#3.