case/spend: Allow uploading a receipt to case spends
This change allows uploading a receipt file (which is not limited on
file type) associated with a case spend. A receipt can be uploaded on
creation of the case spend, or on the spend edit screen.

If the spend is marked as a reimbursement and a receipt is not uploaded
on creation of the spend, a task is created automatically, assigned to
the user who created the spend, prompting them to upload a receipt.

Because of the above, this commit makes changes to the task renderable
view and the case view, as well as changing the case task DB table, to
allow for a task to be created by the system (where the task author is
set to null in the database).

The new CaseSpendReceiptController allows for downloading or viewing
the uploaded receipt.

Closes againstprisons/kukupa#7.