system/mailtemplates: Fixes two bugs - missing `.save` and table non-null
The first bug here raised a "ciphertext failed validation" exception,
which was caused by not saving the mail template object before performing
encryption operations. Adding a `.save` to the end of the line creating
the mail template object fixed this.

The second bug was caused by the initial migration for mail templates
setting the `name` and `content` fields to not allow null values, for
which a migration has been added to correct this.