user, case: Add role-based user "tags"
Tags are shown beside user names in various places, such as the list of
users to assign a case or case task to. Tags are granted to users with
roles of the format `tag:<tag_slug>` and the visible names of the tags
are translatable (translation key of the format `tag/<tag_slug>`).

The two default tags are `coordinator` (Advocacy coordinator) and
`advocate` (Advocacy volunteer).

To add a new user tag, add it to the translation file (either in the
application itself for default tags, or in the theme for instance-
specific tags), and then grant the tag role to the users who need it.

Tag granting also works with role groups, so you can have a role group
for advocates (granting `tag:advocate`), and then any others that may
be needed.