case: Add initial "project" support
This adds a `type` field to cases, which can be set to "case" (which is
the default), or "project" for project cases. Also adds an `is_private`
field to cases, which determines whether a project case is publicly
visible (which they are by default, a private project is only visible
to users that are explicitly added to it).

The project name is stored in the `first_name` field of the case. Most
editable fields are disabled, the only remaining ones in /case/:id/edit
are the first name, the case summary, and whether the case is private.

Other things disabled for project cases:

- Sending mail
- Editing prison / PRN information
- Editing re:connect link settings
- Triage tasks (creation and reset)
- Case purpose

All project cases are stored with the default case purpose (which is
"advocacy"), but the purpose field can not be edited from the UI, and
it's not shown anywhere.