app: Make ApplicationController#before accept arguments
This is a rather large diff, because we had to change every controller
to accept `*args` on their #before methods.

This change also loads the current user and the current case in
CaseController#before, paving the way for the next commit - blocking
access to any case pages if the case type is invalid.