case: Make edit forms & hiding of parts of case UI reasonably generic
This adds the Case::CASE_TYPES and Case::CASE_FIELDS descriptors.

The Case::CASE_TYPES descriptor lays out what fields to show in the edit
form, and what portions of the case UI to hide, based on the case type
field. The `:fields` key of a case type is a list of Symbols, each
corresponding to the name of a field. Optionally, an entry in that array
can be a Hash with a `:field` key (the name of the field), and some of
the options in that hash will override the default values of the field.
This is used for the "project" case type, in which the translation key
for the `first_name` field is overridden.

The Case::CASE_FIELDS descriptor lays out the available fields for the
case edit form. The keys of the CASE_FIELDS hash are the name of the
field. The `tl_key` key in the field hashes is displayed in the UI,
the `type` field describes the type of the field (the available options
are `text`, `number`, `date`, `select`, `checkbox`, and `editor`), the
optional `required` key in the field hashes marks a field as required,
and the optional `editable_name` field is the name of the field in the
`case_editables` hash passed to the view (but if this is not present,
the field name is used instead).

For `select` type fields, the `select_options` key in the field hash
(which is an array of hashes) is used to generate the list of options.
The `value` key in an option hash is what is submitted in the form,
and the `tl_key` key is what is displayed to the user.

I hope that this commit message serves as a good enough description of
how this all works!