dashboard: Display cases assigned to me with new mail
This uses the new `Case#last_updated` and `Case#new_mail?` methods.

The `Case#last_updated` method collates the dates of the latest case
subentries (notes, tasks, spends, mail) and returns the newest.

The `Case#new_mail?` method gets the date of the latest piece of
incoming correspondence (not `sent_by_us`), and compares it to the
`Case#last_updated` output.