case/correspondence: Add fields for email replies, update renderable
This adds the `correspondence_type`, `target_email`, and `email_messageid`
fields to CaseCorrespondence. The usage of these fields is described in
issue againstprisons/kukupa#15.

Also updates the CaseCorrespondence renderable to show the value of the
`target_email` field in the renderable header if the `correspondence_type`
field is set to `email`, which it will be for incoming/outgoing emails.