config: Add `email-outgoing-reply-to`
The value of this field will be used for incoming and outgoing emails.
The value must contain the string `%IDENTIFIER%`, which is gsubbed out
with the case's `email_identifier`.

For outgoing emails, the value of this field (after substitution) is
used as the value of the `Reply-To` header.

For incoming emails, the value of this field (after substitution) is
used for identifying which emails belong to which case, by comparing
against the address portion of the incoming email's `From` header.