case/correspondence: Add sending emails to a given address
Implements the "sending" part of againstprisons/kukupa#15.

This changes the CaseCorrespondenceSendController to accept an `email`
query/form parameter, which if present, will send the email as composed
to that email address. The subject field in this mode is passed directly
through to the outgoing email.

This also enables the email address inclusion part of the mail template
selection, as added in a previous commit.

Outgoing emails composed this way are stored as local files, with the
resulting CaseCorrespondence entry having the `correspondence_type`
field set to 'email', and the `target_email` field set to the email
address that the message was sent to.

Outgoing emails composed this way are rendered using the template found
in `app/views/email_templates/reply_layout.{html,txt}.erb`, which has
the case's `email_identifier` passed to it. That `email_identifier` is
included in the template as part of a "Please keep your reply above
this line" box at the top of the email, which will allow for the future
part of the implementation of againstprisons/kukupa#15 to scan for that
line of text, and trim everything below it before storing the content.

This has been tested as much as I can test it without configuring my
development instance to use a real mail server. I'm pretty confident
that it works as intended, but we'll see.