case/correspondence: Initial pass of incoming mail import
The `CaseCorrespondence.new_from_incoming_email` method takes a
`Mail::Message` instance.

It matches the `To` address in the message against a Regexp constructed
from the `email-outgoing-reply-to` config value to retrieve the case
email identifier, returning early if it couldn't match.

It gets the HTML part of the message, and if there is no HTML part of
the message, it treats the text part as Markdown, renders it to HTML,
and uses that instead.

It checks if the "please reply above this line" banner exists within the
message - if that's present, it takes everything above that banner, else
it takes the entire email content. It then runs a Sanitize pass on the
email content.

Finally, it saves the email content HTML as a local file, constructs a
CaseCorrespondence entry with that file, and returns that entry.

This still needs a bit of work, especially regarding trimming the email
content, but it works as a proof of concept. This implements part of the
email receiving logic required for againstprisons/kukupa#15.