case/correspondence: Add SyncIncomingEmailWorker
This worker connects to the IMAP server defined in the `email-imap-host`
configuration entry, retrieves the latest 20 emails, and feeds those
messages to `CaseCorrespondence.new_from_incoming_email`.

Messages that parse successfully (the `new_from_incoming_email` call
returns a CaseCorrespondence entry) are deleted from the IMAP server.
Messages that do _not_ parse successfully are left where they are.

This has been tested with the PAPA email server, and it works as it
should.