case: Allow creating a re:connect penpal and/or advocacy relationship
This adds the Case#create_in_reconnect! method, which will create a
penpal in re:connect for the case (using the case's PRN as the unique
identifier), and a relationship with the re:connect penpal defined in
the `reconnect-penpal-id` configuration key.

The penpal creation API call can be skipped by passing the option
`:skip_penpal_creation => true`, and the relationship creation API call
can be similarly skipped with `:skip_relationship_creation => true`.

This method will be called by the outside request form on case creation
(since we will have a valid prison in that case), by a regular worker
(which will only run for cases with an empty `reconnect_id` and a non-
empty `prisoner_number` and `prison`), and will be able to be called
manually on the case edit page (by users with the permission to change
re:connect links).

Whether this method does anything can be toggled with the configuration
key `reconnect-create-penpals`.