js/usersearch: Style update, add search help
Selectable items (users in search results, the "search again" button)
are now styled similar to the rest of the buttons in the interface,
so that they appear "clickable" at a glance.

This also adds the <SearchHelp> component to the user search JS, with
a small description of how to search for users with the widget.