case/correspondence: Add automatic task on new prisoner correspondence
This allows for designating a list of users (via a configuration key)
per prison, so that when new incoming prisoner mail is received, a user
is randomly chosen from the list assigned to that prison, and a task is
assigned to them.

Configuration key "correspondence-task-prison-users" is a JSON hash of
`String -> Array<UserId>` where the key is the prison ID as a string,
and the value is an array of user IDs to choose from when assigning a
task. If a key is not present for a given prison, or the array for that
prison is empty, a task is not created.