models/caseunassignednewrequest: initialize
This adds a migration and model class for case_unassigned_new_requests,
a storage table (which will have it's contents refreshed by a background
worker) to indicate which unassigned cases have a new "request" (defined
as either new prisoner correspondence, or a new outside request).

The contents of this table will be displayed on the dashboard for users
with the `case:view_all` permission.